adjusting methods to cli template
